refactor(plugin): rename containermanager to pluginruntimehandler
